+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.3] - 2020-12-01
+### Added
+- Install fwcloud-websrv.
+- New directory structure:
+    /opt/fwcloud/websrv
+    /opt/fwcloud/ui
+    /opt/fwcloud/api
+    /opt/fwcloud/updater
+
 ## [1.0.2] - 2020-11-24
 ### Added
 - Steps required for install fwcloud-updater.